Added rectangles and characters
I have a few different cases for rectangle drawing, one of which is
optimized for large rectangles.  Character drawing is a bit strange,
and the letters have to be stored upside-down.  This is still
primarily a port of libbone to Go, though.  Strings are next, and
with Go's string type it'll be a lot nicer than C.